fix: resolve narrowing conversion errors in main.cpp
- Add static_cast<float>() for size_t * float operations
- Add static_cast<int>() for float / float to int conversions
- Fix syntax error from sed script (extra parenthesis)

+++ apps/ui-sandbox/main.cpp
@@ -68,7 +68,7 @@ void MouseButtonCallback(GLFWwindow* window, int button, int action, int mods) {
 
 		// Check which scene was clicked
 		if (clickY >= kMenuY + kHeaderHeight) {
-			int clickedIndex = static_cast<int>(static_cast<int>((clickY - kMenuY - kHeaderHeight) / kLineHeight));
+			int clickedIndex = static_cast<int>((clickY - kMenuY - kHeaderHeight) / kLineHeight);
 			if (clickedIndex >= 0 && clickedIndex < static_cast<int>(g_menuState.m_sceneNames.size())) {
 				// Switch to selected scene
 				engine::SceneManager::Get().SwitchTo(g_menuState.m_sceneNames[clickedIndex]);
